requirements.txt
numpy>=1.24.0
sympy>=1.12
model2vec>=0.2.0
faiss-cpu>=1.7.4  # For efficient similarity search